rfmp-web/requirements.txt
# Web UI Server Requirements
flask==3.0.0
flask-cors==4.0.0
flask-sock>=0.7.0
gevent>=23.0.0
gevent-websocket>=0.10.1
gunicorn>=21.0.0
python-dotenv==1.0.0
requests>=2.31.0
simple-websocket>=1.0.0
websocket-client>=1.6.0